Create F. Digits Summation.cpp
The (%) is an arithmetic operator known as modulus. It computes the remainder of the integer division. By dividing normally or implementing other forms may lead the code to remain as a bug i.e we are unable to compile the code to execute here suitably, so we used the (MOD) operator so that we can find the summation between the 2nd digit of the input that we have taken respectively.