Create P - First digit !.cpp
int qempty()
{
    return (f == r ? 1 : 0);
}

The question mark is the conditional operator.
 For example the above  code could be rewritten as
{
  if(f==r)
    return 1;
  else
    return 0;
}
We have written the code with both methods.
Some people refer to ?: it as "the ternary operator", because it is the only ternary operator (i.e. operator that takes three arguments) in the language they are using.
except, instead of statements you have expressions; so you can use it in the middle of a more complex statement.

It's useful for writing succinct code, but can be overused to create hard to maintain code.