Create V - Comparison.cpp
What is meant by the ternary operator?
The ternary operator is an operator that exists in some programming languages, which takes three operands rather than the typical one or two that most operators use
Here we used the ternary operator (? and :) .These are used instead of if and else statement so that we can take more than one condition and the code looks fresh and not disorganised. It is also used to make our code shorter and conclude time reduction.